requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.4.26
distro==1.9.0
h11==0.14.0
httpcore==1.0.3
httpx==0.27.0
idna==3.10
openai==1.3.7
pinecone-client==3.0.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
